Drive forms remove DP + finished antiform

+++ src/main/java/wehavecookies56/kk/api/driveforms/DriveForm.java
@@ -10,7 +10,7 @@ public abstract class DriveForm {
 
 	public abstract ResourceLocation getTexture();
 
-	public abstract int getCost();
+	public abstract double getCost();
 
 	public abstract void initDrive(EntityPlayer player);
